refactoring(all) 바이너리 분리 및 .service 파일 기반 제어
