+++ src/server/MalwareScan.cpp
