+++ src/server/MalwareScan.h
