+++ src/server/RsyslogManager.cpp
