+++ src/server/RsyslogRule.cpp
