+++ src/server/RsyslogRule.h
