+++ src/server/ScheduledScan.cpp
