+++ src/server/ScheduledScanWatcher.h
