+++ src/shared/Paths.h
